Replaced size_t wiht ints

+++ src/entities/building.h
@@ -28,7 +28,7 @@ typedef struct _building Building;
  * as an argument, then sets every other value according to the established data
  * If it fails, NULL is returned.
  */
-Building *building_new(size_t id);
+Building *building_new(int id);
 
 /*
  * Frees memory allocated by building_new, it will only free it if the pointer
@@ -46,7 +46,7 @@ Building *building_level_up(Building *);
  * Associates a building with the id of a tile.
  * Returns a NULL value if an error occurs.
  */
-Building *building_place(Building *, size_t);
+Building *building_place(Building *, int);
 
 /*
  * Adds or subtract depending on the increment. If the value were going to be
@@ -57,16 +57,16 @@ Building *building_edit_health(Building *, int);
 
 /*GETTERS*/
 
-size_t building_get_id(Building *);
+int building_get_id(Building *);
 
-size_t building_get_level(Building *);
+int building_get_level(Building *);
 
-size_t building_get_unlocking_level(Building *);
+int building_get_unlocking_level(Building *);
 
-size_t building_get_health(Building *);
+int building_get_health(Building *);
 
-size_t building_get_cost(Building *);
+int building_get_cost(Building *);
 
-size_t building_get_tile_id(Building *);
+int building_get_tile_id(Building *);
 
-size_t building_get_base_resources(Building *);
+int building_get_base_resources(Building *);
